+++ lib/features/contraction/presnetation/view/widgets/contract_success_view_widgets/contract_success_view_body.dart
@@ -5,9 +5,12 @@ import 'package:test1/core/constants/vertical_and_horizontal_space.dart';
 import 'package:test1/core/theming/styles.dart';
 import 'package:test1/core/widgets/custom_button.dart';
 import 'package:pinput/pinput.dart';
+import 'package:test1/core/widgets/show_alert_dialog.dart';
+import 'package:test1/features/contraction/presnetation/view/widgets/contract_success_view_widgets/wallet_balance.dart';
 
 import '../../../../../../core/constants/methods.dart';
 import '../../../../../../core/theming/my_colors.dart';
+import 'contract_success_alert_dialog_content.dart';
 
 class ContractSuccessViewBody extends StatefulWidget {
   const ContractSuccessViewBody({super.key});
@@ -50,7 +53,7 @@ class _ContractSuccessViewBodyState extends State<ContractSuccessViewBody> {
                     .copyWith(color: MyColors.kPrimaryColor),
               ),
               const VerticalSpacer(16),
-               WalletBalanceCard(),
+              WalletBalanceCard(),
               const VerticalSpacer(37),
               SizedBox(
                 child: CustomButton(
@@ -58,7 +61,12 @@ class _ContractSuccessViewBodyState extends State<ContractSuccessViewBody> {
                       .copyWith(color: Colors.white),
                   text: 'ادفع ‏9,800.00 لتفعيل التعاقد',
                   backGroundColor: MyColors.kPrimaryColor,
-                  onPressed: () {},
+                  onPressed: () {
+                    showAlertDialog(
+                      context,
+                      const ContractSuccessAlertDialogContent(),
+                    );
+                  },
                   borderRadius: BorderRadius.circular(8),
                 ),
               ),
@@ -88,64 +96,4 @@ class _ContractSuccessViewBodyState extends State<ContractSuccessViewBody> {
   }
 }
 
-class WalletBalanceCard extends StatelessWidget {
-   TextEditingController walletController= TextEditingController();
-   WalletBalanceCard({
-    super.key,
-  });
 
-  @override
-  Widget build(BuildContext context) {
-    return Container(
-      padding: const EdgeInsets.symmetric(vertical: 10, horizontal: 16),
-      decoration: BoxDecoration(
-        color: Colors.grey[300],
-        borderRadius: BorderRadius.circular(10),
-      ),
-      height: 55.h,
-      child: Row(
-        mainAxisAlignment: MainAxisAlignment.center,
-        children: [
-          SizedBox(
-            width: getWidth(context) * 0.35,
-            child: Text(
-              'لديك رصيد محفظة\n60 ريال يمكنك استخدامها',
-              style: MyTextStyles.font12Weight500,
-            ),
-          ),
-          const HorizontalSpacer(16),
-          Pinput(
-            length: 1,
-            mainAxisAlignment: MainAxisAlignment.center,
-            controller: walletController,
-            onChanged: (v){
-              walletController.text=v;
-            },
-            defaultPinTheme: PinTheme(
-              width: 50.w,
-              height: 32.h,
-              textStyle: MyTextStyles.font12Weight500,
-              decoration: BoxDecoration(
-                color: Colors.white,
-                borderRadius: BorderRadius.circular(7),
-                border: Border.all(color: Colors.grey),
-              ),
-            ),
-          ),
-          const Spacer(),
-          SizedBox(
-            width: 66.w,
-            child: CustomButton(
-              textStyle:
-                  MyTextStyles.font12Weight500.copyWith(color: Colors.white),
-              text: 'استبدال',
-              backGroundColor: MyColors.kPrimaryColor,
-              onPressed: () {},
-              borderRadius: BorderRadius.circular(5),
-            ),
-          ),
-        ],
-      ),
-    );
-  }
-}
